requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.3
google-auth==2.17.2
google-auth-oauthlib==1.0.0
mysql-connector-python==8.0.32
oauthlib==3.2.2
requests==2.28.2
requests-oauthlib==1.3.1
SQLAlchemy==2.0.9
twilio==8.0.0
